+++ Core/Utilities/Maths/Maths.cpp
@@ -4,7 +4,7 @@ namespace Venus { namespace Maths {
 
 float Math::toRadians(float degrees)
 {
-	return degrees * (M_PI / 180.0f);
+	return (float)(degrees * (M_PI / 180.0f));
 }
 
 float Math::Tan(float radians)
